completed hw06, tictactoe

+++ src/homework/06_tic_tac_toe/tic_tac_toe.cpp
@@ -6,7 +6,6 @@ using std::cout;
 bool TicTacToe::game_over()
 {
     return check_board_full();
-
 }
 
 void TicTacToe::start_game(std::string first_player)
@@ -17,13 +16,13 @@ void TicTacToe::start_game(std::string first_player)
 
 void TicTacToe::mark_board(int position)
 {
-    pegs[position -1] = player;
+    pegs[position - 1] = player;
     set_next_player();
 }
 
 void TicTacToe::display_board() const
 {
-    for (int i=0; i < pegs.size(); i+=3)
+    for(long unsigned int i=0; i < pegs.size(); i+=3)
     {
         cout<<pegs[i]<<"|"<<pegs[i+1]<<"|"<<pegs[i+2]<<"\n";
     }
@@ -35,7 +34,7 @@ void TicTacToe::set_next_player()
     {
         player = "O";
     }
-    else
+    else 
     {
         player = "X";
     }
@@ -45,7 +44,7 @@ bool TicTacToe::check_board_full()
 {
     auto full = true;
 
-    for(auto &peg : pegs)
+    for(auto &peg: pegs)
     {
         if(peg == " ")
         {
@@ -53,12 +52,13 @@ bool TicTacToe::check_board_full()
             break;
         }
     }
+
     return full;
 }
 
 void TicTacToe::clear_board()
 {
-    for (auto &peg: pegs)
+    for(auto &peg: pegs)
     {
         peg = " ";
     }
